requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
urllib3==1.26.7
validator==0.7.1
Werkzeug==2.0.2
